# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Playlisted API
  version: 0.2.0
  summary: OpenAPI-first REST API for the Playlisted audio platform.
  description: |
    Contract for the Playlisted backend. The server validates requests and responses
    against this document, and the client SDK is generated from the same source.
servers:
  - url: http://localhost:4000
    description: Local development
security: []
...
paths:
...
  /api/v1/auth/me:
    get:
      tags: [Auth]
      operationId: getMe
      summary: Get the current authenticated user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/me/playlists:
    get:
      tags: [Me]
      operationId: listMyPlaylists
      summary: List playlists owned by the current user
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: Owned playlists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlaylistListResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/me/recordings:
    get:
      tags: [Me]
      operationId: listMyRecordings
      summary: List recordings uploaded by the current user
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: Owned recordings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecordingListResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/me/analytics/recordings:
    get:
      tags: [Analytics]
      operationId: getAnalyticsRecordings
      summary: Get per-track analytics for the current creator
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: range
          schema:
            $ref: '#/components/schemas/ChartRange'
        - in: query
          name: sortBy
          schema:
            type: string
            enum: [plays, duration, completion]
            default: plays
        - in: query
          name: order
          schema:
            type: string
            enum: [asc, desc]
            default: desc
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: Per-track analytics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalyticsRecordingsResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/users:
    get:
      tags: [Users]
      operationId: listUsers
      summary: List users
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - in: query
          name: role
          schema:
            $ref: '#/components/schemas/UserRole'
        - in: query
          name: status
          schema:
            $ref: '#/components/schemas/UserStatus'
        - in: query
          name: featured
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: User collection
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserListResponse'
    post:
      tags: [Users]
      operationId: createUser
      summary: Create a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDetail'
        '409':
          description: User already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/users/me:
    patch:
      tags: [Users]
      operationId: updateMyProfile
      summary: Update the current user's profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
      responses:
        '200':
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDetail'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Username taken
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/playlists:
    get:
      tags: [Playlists]
      operationId: listPlaylists
      summary: List playlists
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - in: query
          name: ownerId
          schema:
            type: string
        - in: query
          name: visibility
          schema:
            $ref: '#/components/schemas/Visibility'
        - in: query
          name: status
          schema:
            $ref: '#/components/schemas/PublishStatus'
      responses:
        '200':
          description: Playlist collection
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlaylistListResponse'
    post:
      tags: [Playlists]
      operationId: createPlaylist
      summary: Create a playlist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePlaylistRequest'
      responses:
        '201':
          description: Playlist created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlaylistDetail'
        '404':
          description: Owner not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Slug already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/recordings:
    get:
      tags: [Recordings]
      operationId: listRecordings
      summary: List recordings
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - in: query
          name: uploaderId
          schema:
            type: string
        - in: query
          name: playlistId
          schema:
            type: string
        - in: query
          name: recordingType
          schema:
            $ref: '#/components/schemas/RecordingType'
        - in: query
          name: visibility
          schema:
            $ref: '#/components/schemas/Visibility'
        - in: query
          name: status
          schema:
            $ref: '#/components/schemas/PublishStatus'
      responses:
        '200':
          description: Recording collection
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecordingListResponse'
    post:
      tags: [Recordings]
      operationId: createRecording
      summary: Create a recording and attach it to its canonical playlist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRecordingRequest'
      responses:
        '201':
          description: Recording created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecordingDetail'
        '404':
          description: Uploader or playlist not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Recording conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/admin/tags:
    get:
      tags: [Admin]
      operationId: adminListTags
      summary: List all tags (admin)
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: kind
          schema:
            $ref: '#/components/schemas/TagKind'
      responses:
        '200':
          description: Tag list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminTagListResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags: [Admin]
      operationId: adminCreateTag
      summary: Create a tag (admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminCreateTagRequest'
      responses:
        '201':
          description: Created tag
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminTag'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/admin/homepage-features:
    get:
      tags: [Admin]
      operationId: adminListHomepageFeatures
      summary: List all homepage features including inactive (admin)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Feature list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminHomepageFeatureListResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags: [Admin]
      operationId: adminCreateHomepageFeature
      summary: Create a homepage feature (admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminCreateHomepageFeatureRequest'
      responses:
        '201':
          description: Created feature
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminHomepageFeature'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/admin/users:
    get:
      tags: [Admin]
      operationId: adminListUsers
      summary: List all users with full details including email (admin)
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - in: query
          name: role
          schema:
            $ref: '#/components/schemas/UserRole'
        - in: query
          name: status
          schema:
            $ref: '#/components/schemas/UserStatus'
        - in: query
          name: q
          schema:
            type: string
      responses:
        '200':
          description: User list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserListResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
  parameters:
    Page:
      in: query
      name: page
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      in: query
      name: pageSize
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  schemas:
    UserRole:
      type: string
      enum: [LISTENER, CREATOR, EDITOR, ADMIN]
    UserStatus:
      type: string
      enum: [ACTIVE, SUSPENDED, INVITED]
    PlaylistType:
      type: string
      enum: [PLAYLIST, ALBUM, MIX, PODCAST_CHANNEL, RELEASE]
    RecordingType:
      type: string
      enum: [SONG, PODCAST_EPISODE, MIX_SEGMENT, OTHER]
    Visibility:
      type: string
      enum: [PUBLIC, UNLISTED, PRIVATE]
    PublishStatus:
      type: string
      enum: [DRAFT, PUBLISHED, ARCHIVED]
    TagKind:
      type: string
      enum: [GENRE, MOOD, CATEGORY, SCENE]
...
    AuthUser:
      type: object
      additionalProperties: false
      required:
        [id, email, username, displayName, role, status, isFeaturedArtist, createdAt, updatedAt]
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        username:
          type: string
        displayName:
          type: string
        bio:
          type: [string, 'null']
        avatarUrl:
          type: [string, 'null']
          format: uri-reference
        heroImageUrl:
          type: [string, 'null']
          format: uri-reference
        role:
          $ref: '#/components/schemas/UserRole'
        status:
          $ref: '#/components/schemas/UserStatus'
        isFeaturedArtist:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    MeResponse:
      type: object
      additionalProperties: false
      required: [user]
      properties:
        user:
          $ref: '#/components/schemas/AuthUser'
    ErrorResponse:
      type: object
      additionalProperties: true
      required: [error, message]
      properties:
        error:
          type: string
        message:
          type: string
        details: {}
    PaginationMeta:
      type: object
      additionalProperties: false
      required: [page, pageSize, total]
      properties:
        page:
          type: integer
        pageSize:
          type: integer
        total:
          type: integer
...
    UserSummary:
      type: object
      required:
        [id, email, username, displayName, role, status, isFeaturedArtist, createdAt, updatedAt]
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        username:
          type: string
        displayName:
          type: string
        bio:
          type: [string, 'null']
        avatarUrl:
          type: [string, 'null']
          format: uri-reference
        heroImageUrl:
          type: [string, 'null']
          format: uri-reference
        role:
          $ref: '#/components/schemas/UserRole'
        status:
          $ref: '#/components/schemas/UserStatus'
        isFeaturedArtist:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    UserPlaylistSummary:
      type: object
      additionalProperties: false
      required:
        [id, href, ownerId, title, slug, type, visibility, status, featured, isPinnedOnProfile, itemCount, totalDurationSeconds, createdAt, updatedAt]
      properties:
        id:
          type: string
        href:
          type: string
        ownerId:
          type: string
        title:
          type: string
        slug:
          type: string
        description:
          type: [string, 'null']
        coverArtUrl:
          type: [string, 'null']
          format: uri-reference
        type:
          $ref: '#/components/schemas/PlaylistType'
        visibility:
          $ref: '#/components/schemas/Visibility'
        status:
          $ref: '#/components/schemas/PublishStatus'
        featured:
          type: boolean
        isPinnedOnProfile:
          type: boolean
        itemCount:
          type: integer
        totalDurationSeconds:
          type: integer
        publishedAt:
          type: [string, 'null']
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    UserDetail:
      allOf:
        - $ref: '#/components/schemas/UserSummary'
        - type: object
          additionalProperties: false
          required: [publicPlaylists]
          properties:
            publicPlaylists:
              type: array
              items:
                $ref: '#/components/schemas/UserPlaylistSummary'
    UserListResponse:
      type: object
      additionalProperties: false
      required: [data, meta]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/UserSummary'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    CreateUserRequest:
      type: object
      additionalProperties: false
      required: [email, displayName]
      properties:
        email:
          type: string
          format: email
        username:
          type: string
          minLength: 1
          maxLength: 50
        displayName:
          type: string
          minLength: 1
          maxLength: 120
        passwordHash:
          type: [string, 'null']
        bio:
          type: [string, 'null']
        avatarUrl:
          type: [string, 'null']
          format: uri-reference
        heroImageUrl:
          type: [string, 'null']
          format: uri-reference
        role:
          $ref: '#/components/schemas/UserRole'
        status:
          $ref: '#/components/schemas/UserStatus'
        isFeaturedArtist:
          type: boolean
          default: false
...
    OwnerSummary:
      type: object
      additionalProperties: false
      required: [id, username, displayName, role]
      properties:
        id:
          type: string
        username:
          type: string
        displayName:
          type: string
        avatarUrl:
          type: [string, 'null']
          format: uri-reference
        heroImageUrl:
          type: [string, 'null']
          format: uri-reference
        role:
          $ref: '#/components/schemas/UserRole'
    RecordingSummary:
      type: object
      additionalProperties: false
      required:
        [id, uploaderId, publishedPlaylistId, title, audioUrl, recordingType, visibility, status, explicit, playCount, createdAt, updatedAt]
      properties:
        id:
          type: string
        uploaderId:
          type: string
        publishedPlaylistId:
          type: string
        title:
          type: string
        description:
          type: [string, 'null']
        audioUrl:
          type: string
          format: uri-reference
        audioMimeType:
          type: [string, 'null']
        audioBytes:
          type: [integer, 'null']
        durationSeconds:
          type: [integer, 'null']
        artworkUrl:
          type: [string, 'null']
          format: uri-reference
        recordingType:
          $ref: '#/components/schemas/RecordingType'
        visibility:
          $ref: '#/components/schemas/Visibility'
        status:
          $ref: '#/components/schemas/PublishStatus'
        trackNumber:
          type: [integer, 'null']
        episodeNumber:
          type: [integer, 'null']
        explicit:
          type: boolean
        releaseDate:
          type: [string, 'null']
          format: date-time
        publishedAt:
          type: [string, 'null']
          format: date-time
        playCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    RecordingDetail:
      allOf:
        - $ref: '#/components/schemas/RecordingSummary'
        - type: object
          additionalProperties: false
          required: [uploader, publishedPlaylist]
          properties:
            uploader:
              $ref: '#/components/schemas/OwnerSummary'
            publishedPlaylist:
              $ref: '#/components/schemas/PublishedPlaylistReference'
    RecordingListResponse:
      type: object
      additionalProperties: false
      required: [data, meta]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/RecordingSummary'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    CreateRecordingRequest:
      type: object
      additionalProperties: false
      required: [uploaderId, publishedPlaylistId, title, audioUrl]
      properties:
        uploaderId:
          type: string
        publishedPlaylistId:
          type: string
        title:
          type: string
          minLength: 1
          maxLength: 191
        description:
          type: [string, 'null']
        audioUrl:
          type: string
          format: uri-reference
        audioMimeType:
          type: [string, 'null']
        audioBytes:
          type: [integer, 'null']
          minimum: 0
        durationSeconds:
          type: [integer, 'null']
          minimum: 0
        artworkUrl:
          type: [string, 'null']
          format: uri-reference
        recordingType:
          $ref: '#/components/schemas/RecordingType'
        visibility:
          $ref: '#/components/schemas/Visibility'
        status:
          $ref: '#/components/schemas/PublishStatus'
        trackNumber:
          type: [integer, 'null']
          minimum: 0
        episodeNumber:
          type: [integer, 'null']
          minimum: 0
        explicit:
          type: boolean
          default: false
        releaseDate:
          type: [string, 'null']
          format: date-time
        publishedAt:
          type: [string, 'null']
          format: date-time
    PlaylistSummary:
      type: object
      required:
        [id, href, ownerId, title, slug, type, visibility, status, featured, isPinnedOnProfile, itemCount, totalDurationSeconds, createdAt, updatedAt, owner]
      properties:
        id:
          type: string
        href:
          type: string
        ownerId:
          type: string
        title:
          type: string
        slug:
          type: string
        description:
          type: [string, 'null']
        coverArtUrl:
          type: [string, 'null']
          format: uri-reference
        type:
          $ref: '#/components/schemas/PlaylistType'
        visibility:
          $ref: '#/components/schemas/Visibility'
        status:
          $ref: '#/components/schemas/PublishStatus'
        featured:
          type: boolean
        isPinnedOnProfile:
          type: boolean
        itemCount:
          type: integer
        totalDurationSeconds:
          type: integer
        publishedAt:
          type: [string, 'null']
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        owner:
          $ref: '#/components/schemas/OwnerSummary'
        tags:
          type: array
          items:
            $ref: '#/components/schemas/Tag'
    PlaylistDetail:
      type: object
      additionalProperties: false
      required:
        [id, href, ownerId, title, slug, type, visibility, status, featured, isPinnedOnProfile, itemCount, totalDurationSeconds, createdAt, updatedAt, owner, recordings]
      properties:
        id:
          type: string
        href:
          type: string
        ownerId:
          type: string
        title:
          type: string
        slug:
          type: string
        description:
          type: [string, 'null']
        coverArtUrl:
          type: [string, 'null']
          format: uri-reference
        type:
          $ref: '#/components/schemas/PlaylistType'
        visibility:
          $ref: '#/components/schemas/Visibility'
        status:
          $ref: '#/components/schemas/PublishStatus'
        featured:
          type: boolean
        isPinnedOnProfile:
          type: boolean
        itemCount:
          type: integer
        totalDurationSeconds:
          type: integer
        publishedAt:
          type: [string, 'null']
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        owner:
          $ref: '#/components/schemas/OwnerSummary'
        tags:
          type: array
          items:
            $ref: '#/components/schemas/Tag'
        recordings:
          type: array
          items:
            $ref: '#/components/schemas/RecordingInPlaylist'
    PlaylistListResponse:
      type: object
      additionalProperties: false
      required: [data, meta]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/PlaylistSummary'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    CreatePlaylistRequest:
      type: object
      additionalProperties: false
      required: [ownerId, title]
      properties:
        ownerId:
          type: string
        title:
          type: string
          minLength: 1
          maxLength: 191
        description:
          type: [string, 'null']
        coverArtUrl:
          type: [string, 'null']
          format: uri-reference
        type:
          $ref: '#/components/schemas/PlaylistType'
        visibility:
          $ref: '#/components/schemas/Visibility'
        status:
          $ref: '#/components/schemas/PublishStatus'
        publishedAt:
          type: [string, 'null']
          format: date-time
    UpdateProfileRequest:
      type: object
      additionalProperties: false
      properties:
        displayName:
          type: string
          minLength: 1
          maxLength: 120
        username:
          type: string
          minLength: 1
          maxLength: 50
        bio:
          type: [string, 'null']
        avatarUrl:
          type: [string, 'null']
          format: uri-reference
        heroImageUrl:
          type: [string, 'null']
          format: uri-reference
...
    ChartRange:
      type: string
      enum: [7d, 30d, all]
      default: 30d
...
    AnalyticsRecordingItem:
      type: object
      additionalProperties: false
      required: [recordingId, title, totalPlays, totalPlaySeconds, completionRate]
      properties:
        recordingId:
          type: string
        title:
          type: string
        artworkUrl:
          type: [string, 'null']
          format: uri-reference
        durationSeconds:
          type: [integer, 'null']
        totalPlays:
          type: integer
        totalPlaySeconds:
          type: integer
        completionRate:
          type: number
    AnalyticsRecordingsResponse:
      type: object
      additionalProperties: false
      required: [data, meta]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/AnalyticsRecordingItem'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    AdminTag:
      type: object
      additionalProperties: false
      required: [id, name, slug, kind, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
        kind:
          $ref: '#/components/schemas/TagKind'
        createdAt:
          type: string
          format: date-time
    AdminTagListResponse:
      type: object
      additionalProperties: false
      required: [data]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/AdminTag'
    AdminCreateTagRequest:
      type: object
      additionalProperties: false
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 64
        kind:
          $ref: '#/components/schemas/TagKind'
...
    AdminHomepageFeature:
      type: object
      additionalProperties: false
      required: [id, section, position, isActive, createdAt, updatedAt]
      properties:
        id:
          type: string
        section:
          $ref: '#/components/schemas/AdminHomepageSection'
        position:
          type: integer
        titleOverride:
          type: [string, 'null']
        subtitleOverride:
          type: [string, 'null']
        description:
          type: [string, 'null']
        imageUrl:
          type: [string, 'null']
          format: uri-reference
        isActive:
          type: boolean
        startsAt:
          type: [string, 'null']
          format: date-time
        endsAt:
          type: [string, 'null']
          format: date-time
        playlistId:
          type: [string, 'null']
        userId:
          type: [string, 'null']
        editorialPostId:
          type: [string, 'null']
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        playlist:
          oneOf:
            - $ref: '#/components/schemas/AdminHomepageFeatureRef'
            - type: 'null'
        user:
          oneOf:
            - $ref: '#/components/schemas/AdminHomepageUserRef'
            - type: 'null'
    AdminHomepageFeatureListResponse:
      type: object
      additionalProperties: false
      required: [data]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/AdminHomepageFeature'
    AdminCreateHomepageFeatureRequest:
      type: object
      additionalProperties: false
      required: [section]
      properties:
        section:
          $ref: '#/components/schemas/AdminHomepageSection'
        position:
          type: integer
        titleOverride:
          type: [string, 'null']
        subtitleOverride:
          type: [string, 'null']
        description:
          type: [string, 'null']
        imageUrl:
          type: [string, 'null']
          format: uri-reference
        isActive:
          type: boolean
          default: true
        startsAt:
          type: [string, 'null']
          format: date-time
        endsAt:
          type: [string, 'null']
          format: date-time
        playlistId:
          type: [string, 'null']
        userId:
          type: [string, 'null']
...